Rename file, too
